Add helper.deprecated function. Renamed grid to static_grid
