+++ example/gui/main/main.gui
@@ -10961,11 +10961,5 @@ layers {
   name: "text_top"
 }
 material: "/builtins/materials/gui.material"
-layouts {
-  name: "Landscape"
-}
-layouts {
-  name: "Portrait"
-}
 adjust_reference: ADJUST_REFERENCE_PARENT
 max_nodes: 512
